service implementation

+++ LoanApplication/src/main/java/com/salimcankaya/service/implementation/CustomerServiceImpl.java
@@ -1,8 +1,11 @@
 package com.salimcankaya.service.implementation;
 
-import org.springframework.beans.factory.annotation.Autowired;
+import java.time.LocalDate;
+
 import org.springframework.stereotype.Service;
 
+import com.salimcankaya.exception.CustomerNotFoundException;
+import com.salimcankaya.exception.DuplicateTcknException;
 import com.salimcankaya.model.Customer;
 import com.salimcankaya.repository.CustomerRepository;
 import com.salimcankaya.repository.LoanRepository;
@@ -20,7 +23,7 @@ public class CustomerServiceImpl implements CustomerService {
 	
 	private final LoanRepository loanRepo;
 	
-	@Autowired
+	
 	public CustomerServiceImpl(CustomerRepository customerRepo, LoanRepository loanRepo) {
 		
 		this.customerRepo = customerRepo;
@@ -31,35 +34,65 @@ public class CustomerServiceImpl implements CustomerService {
 	@Override
 	public Customer getCustomerByTckn(Long tckn) {
 		
-		return customerRepo.findByTckn(tckn).orElseThrow();
+		return customerRepo.findByTckn(tckn)
+				.orElseThrow(() -> new CustomerNotFoundException("Customer with provided tckn" + tckn + "not found!"));
 	}
 
 	@Override
 	public Customer addCustomer(Customer customer) {
 		
-		return customerRepo.save(customer);
+		if(existByTckn(customer.getTckn())) {
+			
+			throw new DuplicateTcknException();
+		
+		} else {
+			
+			return customerRepo.save(customer);
+		}
 	}
 
 	@Override
 	public Customer updateCustomer(Customer customer) {
 		
-		Customer customerFromDatabase = customerRepo.findById(customer.getTckn()).orElseThrow();
+		if(!existByTckn(customer.getTckn())) {
+			
+			throw new CustomerNotFoundException("Customer with provided tckn" + customer.getTckn() + "not found! Update operation is cancelled...");
 		
-		customerFromDatabase.setTckn(customer.getTckn());
-		customerFromDatabase.setName(customer.getName());
-		customerFromDatabase.setLastName(customer.getLastName());
-		customerFromDatabase.setDateOfBirth(customer.getDateOfBirth());
-		customerFromDatabase.setPhoneNumber(customer.getPhoneNumber());
-		customerFromDatabase.setMonthlySalary(customer.getMonthlySalary());
-		customerFromDatabase.setDeposit(customer.getDeposit());
+		} else {
+			
+			return customerRepo.save(customer);
+		}
 		
-		return customerRepo.save(customer);
 	}
 
 	@Override
 	public boolean deleteCustomerByTckn(Long tckn) {
 		
-		return customerRepo.deleteByTckn(tckn);
+		if(!existByTckn(tckn)) {
+			
+			throw new CustomerNotFoundException("Customer with provided tckn" + tckn + "not found! Delete operation is cancelled...");
+		
+		} else {
+			
+			customerRepo.deleteByTckn(tckn);
+			loanRepo.deleteLoansByCustomer_Tckn(tckn);
+			return true;
+		}
+		
+	}
+
+
+	@Override
+	public boolean existByTckn(Long tckn) {
+		
+		return customerRepo.existsById(tckn);
+	}
+
+
+	@Override
+	public boolean existByDateOfBirth(LocalDate dateOfBirth) {
+		
+		return customerRepo.existsByDateOfBirth(dateOfBirth);
 	}
 
 }
